refactor(generation): split threemfExporter into cohesive modules
